Revise EnvironmentOptions
