+++ qiskit_ibm_runtime/options_models/environment_options.py
@@ -72,3 +72,30 @@ class EnvironmentOptions:
         | None
     ) = None
     """Runtime image used for this job."""
+
+
+@dataclass(config=PRIMITIVES_CONFIG)
+class SamplerEnvironmentOptions:
+    """Options related to the execution environment."""
+
+    log_level: LogLevelType = "WARNING"
+    """logging level to set in the execution environment.
+
+    The valid log levels are: ``DEBUG``, ``INFO``, ``WARNING``, ``ERROR``, and ``CRITICAL``.
+    """
+
+    job_tags: list[str] = Field(default_factory=list)
+    """Tags to be assigned to the job.
+
+    The tags can subsequently be used as a filter in the
+    :meth:`qiskit_ibm_runtime.qiskit_runtime_service.jobs()` function call.
+    """
+
+    private: bool = False
+    """Boolean that indicates whether the job is marked as private.
+
+    When set to true, input parameters are not returned, and the results can only be read once.
+    After the job is completed, input parameters are deleted from the service. After the results are
+    read, these are also deleted from the service. When set to false, the input parameters and
+    results follow the standard retention behavior of the API.
+    """
