+++ qiskit_ibm_runtime/options_models/execution_options.py
@@ -12,6 +12,8 @@
 
 """Execution options."""
 
+from typing import Literal
+
 from pydantic.dataclasses import dataclass
 
 from .utils import PRIMITIVES_CONFIG
@@ -33,3 +35,24 @@ class ExecutionOptions:
 
     Default is given by ``backend.default_rep_delay``.
     """
+
+
+@dataclass(config=PRIMITIVES_CONFIG)
+class SamplerExecutionOptions(ExecutionOptions):
+    """Execution options for the sampler primitive.
+
+    Args:
+        init_qubits: Whether to reset the qubits to the ground state for each shot.
+            Inherited from :class:`~.ExecutionOptions`.
+        rep_delay: The repetition delay. Inherited from :class:`~.ExecutionOptions`.
+        meas_type: How to process and return measurement results. This option sets
+            the return type of all classical registers in all sampler pub results.
+
+            * ``"classified"``: Returns a BitArray with classified measurement outcomes.
+            * ``"kerneled"``: Returns complex IQ data points from kerneling the measurement
+              trace, in arbitrary units.
+            * ``"avg_kerneled"``: Returns complex IQ data points averaged over shots,
+              in arbitrary units.
+    """
+
+    meas_type: Literal["classified", "kerneled", "avg_kerneled"] = "classified"
